template/* move to Utility/*

+++ test/template_no_Ruby.test.cpp
@@ -1,12 +1,11 @@
 #define PROBLEM "https://judge.yosupo.jp/problem/many_aplusb"
-#include "./../Utility/Input.cpp"
-#include "./../Utility/Output.cpp"
+#include "./../template.cpp"
 #include <iostream>
 using namespace std;
 
 int main() {
 	for (int t = in; t--;) {
-		long long a = in, b = in;
+		ll a = in, b = in;
 		out(a + b);
 	}
 }
